// Application.WebApp/ClientApp/src/app/interfaces/prejoining/pendingcandidatelist.interface.ts
export interface IPendingCandidateList {
    requisitionDetailId: number;
    candidateId: number;
    candidateNo: string;
    requisitionId: number;
    requisitionNo: string;
    candidateFullName: string;
    verticalId: number;
    verticalName: string;
    designation: number;
    designationName: string;
    functionId: number;
    functionName: string;
    departmentId: number;
    departmentName: string;
    locationId: number;
    locationName: string;
    currentLocationId: number;
    currentLocationName: string;
    gradeId: number;
    gradeName: string;
    docApprovalStatusId: number;
    docApprovalStatus: string;
    reportingManagerDocApprovalStatusId: number;
    reportingManagerDocApprovalStatus: string;
    recruitmentManagerNotApproveDoc: string;
    onboardingManagerDocApprovalStatusId: number;
    onboardingManagerDocApprovalStatus: string;
    onboardingManagerNotApproveDoc: string;
    docApprovalEMPNo: string;
    docApprovalEmpName: string;
    docApprovalDesignation: string;
    candidateOwner: string;
    medicalDocumentCollectionId: number;
    medicalDocumentDoctorApprovalId: number;
    preEmployeeMedicaStatuslId: number;
    preEmployeeMedicalStatus: string;
    candidateBVGReportId: number;
    bvgReportStatus: string;
    dateOfJoining: string;
    joiningDateStatus: string;
    hiringStatusId: number;
    hiringStatusName: string;
    popoverContent: string;
    monthlyGross: number;
    checked: boolean;
    onBoardingManager: number;
}